Exported some JcrUserHomeOAuthTokenStore and required dependencies interfaces and classes
