+++ src/main/java/org/apache/sling/auth/oauth_client/impl/OAuthCallbackServlet.java
@@ -40,8 +40,11 @@ import org.apache.sling.api.SlingHttpServletResponse;
 import org.apache.sling.api.servlets.SlingAllMethodsServlet;
 import org.apache.sling.auth.core.AuthConstants;
 import org.apache.sling.auth.oauth_client.ClientConnection;
+import org.apache.sling.auth.oauth_client.spi.Converter;
 import org.apache.sling.auth.oauth_client.spi.OAuthState;
 import org.apache.sling.auth.oauth_client.spi.OAuthStateManager;
+import org.apache.sling.auth.oauth_client.spi.OAuthTokenStore;
+import org.apache.sling.auth.oauth_client.spi.OAuthTokens;
 import org.apache.sling.servlets.annotations.SlingServletPaths;
 import org.jetbrains.annotations.NotNull;
 import org.osgi.service.component.annotations.Activate;
