+++ src/test/java/org/apache/sling/auth/oauth_client/support/OAuthEnabledSlingServletTest.java
@@ -29,10 +29,10 @@ import org.apache.sling.auth.oauth_client.ClientConnection;
 import org.apache.sling.auth.oauth_client.InMemoryOAuthTokenStore;
 import org.apache.sling.auth.oauth_client.OAuthTokenAccess;
 import org.apache.sling.auth.oauth_client.impl.MockOidcConnection;
-import org.apache.sling.auth.oauth_client.impl.OAuthException;
+import org.apache.sling.auth.oauth_client.spi.OAuthException;
 import org.apache.sling.auth.oauth_client.impl.OAuthTokenRefresher;
-import org.apache.sling.auth.oauth_client.impl.OAuthTokenStore;
-import org.apache.sling.auth.oauth_client.impl.OAuthTokens;
+import org.apache.sling.auth.oauth_client.spi.OAuthTokenStore;
+import org.apache.sling.auth.oauth_client.spi.OAuthTokens;
 import org.apache.sling.auth.oauth_client.impl.TokenAccessImpl;
 import org.apache.sling.testing.mock.sling.junit5.SlingContext;
 import org.apache.sling.testing.mock.sling.junit5.SlingContextExtension;
